Refactor network section in settings
