+++ src/components/settings-menu/NetworkSection/strings.ts
@@ -1,0 +1,4 @@
+export const strings = {
+  update: 'Update',
+  showAllNetworks: 'Show all networks',
+};
