+++ src/components/settings-menu/index.js
@@ -1,6 +1,6 @@
 export { default as CurrencySection } from './CurrencySection';
 export { default as LanguageSection } from './LanguageSection';
-export { default as NetworkSection } from './NetworkSection';
+export { default as NetworkSection } from './NetworkSection/NetworkSection';
 export { default as NotificationsSection } from './NotificationsSection';
 export { default as WalletConnectSessionsSection } from './WalletConnectSessionsSection';
 export { default as SettingsSection } from './SettingsSection';
